[inventory] inventory item manifest
we are basically making a struct that will hold all the information about the item in a manifest, adding it to the component and then when the player picks it up, we make a duplicate that will be used in the inventory later.
The Manifest allows us to carry all of the important Item information along so that the Inventory Item keeps all of that data.